Updated Client UI and VM.
Modified model Client, limited last name, first name and id card number
to 20 characters.
Modified model BoughtProduct, instead of extend Product, use a proprety
Product now.
Completed header for main form, added logout and clear selected client
button.
Added a logo, not used yet.
Added or completed basic UI and VM for Simulation, Dashboard and
Product.
Almost completed Client's UI and VM, completed show/add/modify/choose
client feature.

+++ OKInvestir/Model/TimeInterest.cs
@@ -11,7 +11,7 @@ namespace OKInvestir.Model
     {
         public int Id { get; set; }
         [Required]
-        public decimal Time { get; set; }
+        public int Time { get; set; }
         [Required]
         public decimal Interest { get; set; }
         [Required]
